[utils] E: Add -tmp-dir To Nanvix-Bench For Nanvixd
For a while now nanvixd supports a CLI arg, `-tmp-dir`, to provide a
temporary directory where to store all the used socket files, and other
execution artifacts. These should not leak across executions, but it is
sometimes hard to know if a leak is because of a current bug, or
something that happened in the past.

As a consequence, in this commit I surface this argument all the way to
nanvixd-bench, such that we can set it there, too.

Signed-off-by: Carlos Segarra <carlos@carlossegarra.com>