[libs] B: Make Blocking write_all Truly Blocking
In this commit I address a limitation in the current write_all
implementation of the BlockingSocketStream class, where writes would
sometimes fail with WouldBlock. This can happen, and does happen often,
when we connect a socket in a non-blocking fashion, then convert it to a
blocking socket, and then try to write to it.

In this commit I make sure that the socket is writable when we write to
it, and block otherwise.

Signed-off-by: Carlos Segarra <carlos@carlossegarra.com>