[microvm] F: Adding pause and resume functionality
This implements pausing and resuming via control-plane messages to the VMM via co-design with the kernel. Let's go through a whole pause-resume cycle to understand the implementation.

The changes start at `mod.rs`. The VMM now has communication channels to the MicroVM, which are used for syncing the main thread and the vCPU thread. In the main thread, when the VMM receives a `ControlCommand::PauseMicroVm`, it calls `pause_protocol()` inside `handle_command()`. `pause_protocol()` now calls `MicroVm.asnc_pause()`, then waits for a confirmation from the channel to the MicroVM that the vCPU thread operates. There's plumbing for future timeouts.

Let's now switch to `microvm.rs`. It also has the channels to communicate between the vCPU thread and the main thread. The `async_pause()` function writes to the guest OS's kernel's memory a bit telling it to stop execution. Conceptually, we could say there's a state transition in the MicroVM from "running" to "pausing".

Let's now briefly switch to `lib.rs`. It defines the values that are supposed to be written / read in the kernel memory, and the register's address.

Let's now go deep into the kernel. This runs in the vCPU thread. In `clock.rs`, the `timer_handler()` function runs whenever there's a timer interrupt. It reads the same address that was written to by `async_pause()`. If the value read corresponds to the pause request, it issues a VM exit through PMIO to `DEFAULT_VMM_PORT` with an exit status of `DEFAULT_VMM_PAUSE_CMD`. This status must never be used by other PMIO, or they will collide, since this port is multiplexed.

Let's now go back to `microvm.rs`. Once the VM exit is executed, the vCPU thread goes back to the body of the `MicroVm::run()` function. It's done with `self.vcpu.run()?`, and matches on the `exit_context.reason()`. It matches with `PmioAccess` and goes to `emulator.rs`, then checks the `exit_status` to match the magic number. It sends a message to the VMM confirming the vCPU has stopped. That was the message it was waiting for from the beginning of this explanation. Conceptually, this changes the MicroVM state from "pausing" to "paused". To actually implement being paused, it waits for a message telling it to resume. There's timeout plumbing for this, but I think it's uncalled for, since snapshots may be created at this time, and that takes time.

Now that the MicroVM is paused, we could take a consistent snapshot. However, that's for another PR. Let's turn our attention to resuming the MicroVM.

Back at the VMM (`mod.rs`) in the main thread, we're now handling a `ControlCommand::ResumeMicroVm`. We call `resume_microvm()`, which calls `MicroVm::resume()`. It then sends a message to the channel that was expecting a message to continue the `MicroVm::run()` loop in the vCPU thread. Receiving the message conceptually changes the MicroVM state from `Paused` to `Running`. It's important to say that `resume_microvm()` is synchronous, so its effects will be done before the message is sent, and therefore before it is received. We'll dive deep into this soon.

Let's go to `microvm.rs` and look at `resume_microvm()`. All it does is clear the bit that requested the kernel to pause execution.

Let's focus on ordering. The VMM thread clears the bit, then it sends a message to the vCPU thread telling it to resume execution. This is an asynchronous message, so the VMM may continue handling control commands while the vCPU hasn't really resumed. Suppose the VMM receives a `ControlCommand::Pause` before the vCPU is schedulled. Then, the VMM would execute `pause_protocol()`, and write to the kernel's memory the `PAUSE_REQUESTED` value.  The VMM would eventually stop while waiting for a message from the MicroVM confirming it has paused. At this point, the vCPU thread must be schedulled. It goes back to the `MicroVm::run()` loop, and executes for a quantum. When the interrupt goes off, the kernel reads the memory with the `PAUSE_REQUESTED` value and pauses execution once more. Then, it sends the message to the main thread. Therefore, there are no race conditions in this case, and there's no need to wait for a confirmation message in `Vmm::resume_microvm()`. Just sending a message is enough.