[utils] E: Add -toolchain-bin-dir CLI Arg To Nanvixd
In this commit I extend `nanvixd` with an additional CLI argument,
`-toolchain-bin-dir` to point to the binary directory of the Nanvix
toolchain. This will become necessary with the L2 VM, as we need to know where
is `cloud-hypervisor` installed.

In the process I also clean-up a bit the call signatures of different
functions in `nanvixd` by passing an Arc to the CLI arguments, instead
of unpacking them at the beginning and passing each argument separately.

I only unpack the arguments once we need to initialize the
SandboxConfig, as that is, semantically, something different.

Signed-off-by: Carlos Segarra <carlos@carlossegarra.com>