[utils] E: Nanvixd Assigns Unique Socket Addresses For Each User VM
In this commit I extend nanvixd to assign unique gateway socket
addresses for each user VM. This is easy for non-L2 deployments where we
use UNIX sockets, as we can just add the sandbox ID to the socket
address.

For L2 deployments, however, we need to assign a fresh port inside the
L2 VM. Given that this happens inside the L2, which will be in its own
namespace in the future, collisions are less likely. That being said, we
want to make sure we pick a port range where we don't have overlaps.

Signed-off-by: Carlos Segarra <carlos@carlossegarra.com>