[microvm] E: Adding codepaths to issue pause-resume commands
Added types to every control channel, even though unit types would be enough for some of them. This way, they can be extended cleanly. They all have standardized enumeration names. They're all in the same file, `orchestrator.rs`.

The memory thread and the vCPU threads now have control channels to the main thread.

The `pause_protocol` now includes a step where it sends a message to the memory thread, then it waits for a response from the vCPU thread. Note the memory thread doesn't do anything yet, that's for another PR.

The `resume_microvm` (I guess I should rename it to resume_protocol) now sends a message to the memory thread and waits for a response from the memory thread, then it sends a message to the vCPU thread. Note those messages don't do anything, that's for another PR.