[microvm] F: Implementing pause-resume functionality with kernel co-design
This implements pausing and resuming for the MicroVM only. It doesn't implement it for hyperlight yet.

In `mod.rs`, I implemented two closures that are called by the memory thread when it receives a request to pause. They write to the kernel's memory. Each of them holds a different Arc. I don't find it elegant, but hyperlight wouldn't have an Arc to pass, so I did it for compatibility with hyperlight. These closures write a magic value to a magic register. They are defined in `lib.rs`.

In the kernel, in `clock.rs`, the magic address is read and checked if it matches with the magic value. If it does, `::arch::io::out32()` causes a VM exit and sends another magic value that's also defined in `lib.rs`.

The magic value sent by the kernel is read by the Emulator in `emulator.rs`. It is passed to the MicroVM. Note this value that's passed must not collide with any existing exit code. It 0x3000 is already taken, then we need to change it in `lib.rs` to something else, or change the value when the Emulator passes it to the MicroVM.

In `microvm.rs`, the exit status is checked, and the vCPU thread informs the main thread that is has paused. Then, it loops, waiting for a message to resume. I added some timeout infrastructure, but I'm not sure it makes sense in this case. Resume messages might take arbitrarily long, and that's generally fine. I think timeouts should occur on the other end, the one sending the `Resume` message, rather than the one waiting for a `Resume` message.

Finally, `orchestrator.rs` changed because `VcpuControlResponse::Paused` is now constructed, so it lost its trailing underscore. Since I was already editing the file, I improved its trace messages.

As I mentioned, this does not implement pausing and resuming for hyperlight. To do that, we need to write similar closures / functions that write to the kernel's memory, and add similar behavior to the Emulator / MicroVM, where the exit context is checked and the vCPU thread interacts with the main thread via message passing. Resuming hyperlight might not be possible at this point, but pausing the program  and restarting it should be.