[microvm] E: Removing MicroVM global lock
We need to release the vCPU lock in the middle of `microvm.run()` so the main thread can read the kvm state in the vCPU for snapshots. This removes the global MicroVM lock and adds a lock to the vCPU, so it can be released. In order to do that, another coarse lock was added to the fields which had no locks.