[daemons] B: Refine User VM Spawning To Prevent Race Conditions
In this commit I refine the protocol that linuxd and nanvixd use to
spawn new user VMs in order to avoid having a race between the client
trying to connect to the gateway, and the gateway listener being ready.

To fix this, I make nanvixd first try to connect to the gateway
listener, and only return the user VM id to the client once it has
managed to succesfully connect once.

Signed-off-by: Carlos Segarra <carlos@carlossegarra.com>