[microvm] E: Remove Gateway Structure
In this commit I remove the gateway.rs structure used in the micro VM to
monitor the connection to the system VM.

This commit is a first step towards implementing asynchronous polling
for the micro VM. In particular, it will become necessary to have direct
access to all the streams/queues managed by the I/O thread in order to
multiplex them over a poll.

In addition, the gateway structure was not encapsulating too much
complexity and was a bit misleading in terms of what it encapsulared
(system VM vs Gateway).

Signed-off-by: Carlos Segarra <carlos@carlossegarra.com>