[microvm] B: Properly Catch vCPU Thread Interruption
In this commit I fix a bug regarding the interruption of a vCPU thread.

When we first introduced support to interrupt a vCPU thread using a
signal handler, I incorrectly assumed that the interrupted thread would
return a KVM_EXIT_INTR. However, this is not the case.

Reading the manual more carefully [1] (S4.21)

> Any unblocked signal received (except SIGKILL and SIGSTOP, which retain their
> traditional behaviour) will cause KVM_RUN to return with -EINTR.

So, in fact, we should be checking for an error with errno -EINTR.

[1] https://docs.kernel.org/virt/kvm/api.html

Signed-off-by: Carlos Segarra <carlos@carlossegarra.com>