[microvm] F: Remove Spinning Cores From MicroVM
In this commit I re-architect the micro VM to not spin any cores
needlesly, by having a tread execute a tight loop in a non-blocking
fashion.

Instead, I move the three main components: I/O thread, memory thread,
and orchestrator (i.e. VMM) thread to an asynchronous architecture with
polling.

This means, firstly, moving the main loops of each of these components
to a `poll()` loop. In addition, for the threads monitoring queues, we
need to give each component that may _write_ to those queues, a handle
to a `Waker` structure, in order to notify the `poll` that a message has
been pushed to the queue. It is very important to _drain_ any component
monitored using `poll` before sleeping again in the `poll()` loop, as
otherwise we may lose messages.

Signed-off-by: Carlos Segarra <carlos@carlossegarra.com>