[utils] B: Add Timeout In Nanvixd's Sandbox Clean-Up
In this commit I patch nanvixd to take a timeout when waiting for the
graceful shutdown of the system or user VM.

It is good that we try and cleanly shutdown the different components,
but if something goes wrong nanvixd cannot wait and hang forever.
Instead, if we hit a timeout we send a SIGKILL to the child.

Signed-off-by: Carlos Segarra <carlos@carlossegarra.com>