[libs] B: Fix BlockingSocketStream Read Implementation
In this commit I fix a stall bug in the read implementation for the
BlockingSocketStream class.

In particular, by calling poll before trying to read, we were creating a
situation where we may successfully return a read, but not fully drain
the socket stream. This means that the next time we call poll, we may
not need to call poll at all because the data would already be there.

This bug could manifest in the reader hanging indefinately.

Signed-off-by: Carlos Segarra <carlos@carlossegarra.com>