[build] B: Add User-Vm And Control-Plane APIs To Rlibs
In this commit I add the control-plane-api and user-vm-api libraries to
ALL_HOST_RUST_RLIBS such that they get formatted and linted as part of
our regular make targets.

It seems that after the big refactor, these files were actually already
formatted. In any case this fix is still necessary.

Signed-off-by: Carlos Segarra <carlos@carlossegarra.com>