[scripts] B: Propagate LOG_LEVEL To run-nanvixd.sh
In this commit I fix a bug where the LOG_LEVEL env. var was not honoured
in the nanvixd tests, because `./scripts/run-nanvixd.sh` overwrites
LOG_LEVEL, and expects to receive it via the `-log-level` CLI arg.

This means that we need to modify `./scripts/test-nanvixd.sh`, its
caller, to set the right env. variable.

Signed-off-by: Carlos Segarra <carlos@carlossegarra.com>