[ci] E: Add Echo-Breakdown To CI
In this commit I enable the echo-breakdown-* tests in the CI, and I
include them in the benchmark summary. I think this commit was
accidentally dropped from #1086 when rebasing and merging the PR.

Note that the tming for the L2 VM is slightly off due to a mismatch
between the clock in the host and in the L2 VM. I think there is still
value in keeping it, though.

Signed-off-by: Carlos Segarra <carlos@carlossegarra.com>