[nanvix-bench] E: Re-use Nanvixd In All Benchmarks
In this commit I re-factor the benchmarks in `nanvix-bench` to always
re-use a `nanvixd` instance across all iterations of the same benchmark.
This is in-line with the `nanvix-test` behaviour, and will make it
easier to incorporate elements from `nanvix-test` to `nanvix-bench`.

In the process, I also improve the readability of some methods by
transforming boolean arguments into type-safe alternatives.

Signed-off-by: Carlos Segarra <carlos@carlossegarra.com>