[uservm] B: Snapshot reading correction
Corrects the expected snapshot file structure for loading.

Having 4 different structs inside a file was not helpful. I created a single wrapper structure to make serialization and deserialization straighforward with a single call.

Having a single structure also provides a single source of truth for the ordering inside the file, instead of having to keep the writing and reading orders consistent manually.